Commit Beta Updates
Modified the aiming animation to point straight forward and not tilt up slightly.
Added a bool on the player character "UseAimOffset". This can be used to completely remove the aim offset.
Modified aim offset to only work when aiming using the right mouse button otherwise if just shooting it will aim directly in front of the character.
Added a FireRate variable on the player character that we can use to increase how quickly the weapon can be fired.
Added a function to player character interface to increase/decrease max jump count clamped at min 1 and max 3. This can be used for power ups or debuffs.
Added a function to player character interface to increase/decrease fire rate clamped at min .25 and max 3.
Add a widget to HUD to display the current fire rate.