Commit update AI
Added two new drones, in LevelPrototyping "BP_AI_Drone" and "BP_Drone_01". They work similarly to the spline movement BP you must set the spline and spline settings as well as the desired health.
Drone_01 will do a electric attack that will cause a debuff on the player which slows maxWalk speed for 5 seconds.
Both of these drones are meant to remove the AICharacter, so please remove from levels. This also means we do not need the navMesh anymore.
Changed level 2, to a nighttime theme, but now you will need to add point lights to see.
Added next level functionality to level1 so now transitions to level2.